+++ SalaryComparer/ConsoleLogger.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace SalaryComparer
+{
+    public class ConsoleLogger : ILogger
+    {
+        public void Information(string message)
+        {
+            Console.WriteLine(message);
+        }
+    }
+}