+++ SalaryComparer/ILogger.cs
@@ -1,0 +1,7 @@
+namespace SalaryComparer
+{
+    public interface ILogger
+    {
+        void Information(string message);
+    }
+}